refactor(k8s.construct.api): extract types module
