+++ packages/k8s/construct/api/src/index.ts
@@ -1,1 +1,2 @@
 export * from './api'
+export * from './types'
